clean up nat detection changes
