+++ src/test/resources/testdata/detectnatlist1.txt
